Add files via upload
Measurement_Analysis.sh: This script in bash can be used to do the corresponding tests in modules for the demonstrator. Two kind of tests can be done: Bump connectivity and Tunning module. The first one can show how the good and bad pyxels are distributed in the module chips (good and bad connected) by means of a pyxel matrix. The second one can do a noise test with everyone of the pyxels by means of a three diffirent stages that involves three differents kind of scans.

SD_BOOT-ROOT_FADA_installation_final.sh: With this script, you can boot and root a SD card with ubuntu 22.04. This SD card will be used on a fpga board to make the corresponding tests (measurements and analysis of bump conectivity) of Modules. At the same time, the script will download and install FADA in the SD card.

Module_list.txt: In this list you will find the names of the modules that can be tested with the algorithm. If a new module will be inserted, it must be done before running the code. Aditionally, the list txt file must be by default in the Download folder, if not, its path can be changed in the code.

SDCard_IPAdrr_list.txt: This file contains the names (or codes) of the used SD card IP addresses. If a new one will be used, it must be addes here before running the code.